+++ package.json
@@ -9,11 +9,15 @@
     "dependencies": {
         "bcrypt": "^5.1.0",
         "cookie-parser": "~1.4.4",
+        "csv-parser": "^3.0.0",
+        "csv-writer": "^1.6.0",
         "debug": "~2.6.9",
         "dotenv": "^16.0.3",
         "express": "~4.16.1",
         "joi": "^17.8.3",
+        "json2csv": "^6.0.0-alpha.2",
         "jsonwebtoken": "^9.0.0",
+        "jszip": "^3.10.1",
         "mongoose": "^6.9.2",
         "mongoose-paginate-v2": "^1.7.1",
         "morgan": "~1.9.1",
